requirements.txt
altair==4.2.0
anyio==3.6.2
apturl==0.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
contourpy==1.0.6
cramjam==2.6.2
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.0.4
fastapi==0.88.0
fasteners==0.14.1
fastparquet==2022.12.0
fonttools==4.38.0
fsspec==2022.11.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-language==2.6.1
google-cloud-speech==2.16.2
google-cloud-translate==3.8.4
googleapis-common-protos==1.57.0
googletrans==3.1.0a0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httplib2==0.14.0
httptools==0.5.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
importlib-metadata==5.2.0
importlib-resources==5.10.1
iniconfig==1.1.1
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
monotonic==1.5
netifaces==0.10.4
networkx==2.8.8
nltk==3.8
numpy==1.23.5
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
paramiko==2.6.0
pexpect==4.6.0
phonetics==1.0.5
Pillow==7.0.0
pipreqs==0.4.11
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
proto-plus==1.22.1
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycups==1.9.73
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.2
pytest==7.2.0
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
regex==2022.10.31
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
seaborn==0.12.1
SecretStorage==2.3.1
segtok==1.5.11
semver==2.13.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
starlette==0.22.0
streamlit==1.16.0
systemd-python==234
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.20.0
uvloop==0.17.0
validators==0.20.0
wadllib==1.3.3
watchdog==2.2.0
watchfiles==0.18.1
websockets==10.4
wikipedia==1.4.0
wordcloud==1.8.2.2
xkit==0.0.0
yake==0.4.8
yarg==0.1.9
zipp==3.11.0
